github-merge-pr: handle user using master as pushing repo
Some user may use master or other openwrt repo as the pr repo. This is
problematic as we checkout with the pr repo name and this may conflict
with the current openwrt branch name.

Fix it and checkout the local branch with the pr repo name + the pr
author.

Example:

usertest123:master is checkout to master-usertest123

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>